Añado requisitos funcionales 2.1, 2.2 y 2.3. Falta por implementar de mi sSI el 2.4, 2.5 y 2.6.

+++ include/GestionRecursosEnergeticos.h
@@ -73,7 +73,7 @@ public:
      * @return true si la consulta fue exitosa; false en caso contrario.
      */
     bool consultarIngresosPorInstalacion(const std::string& direccion,
-                                         int& ingresos) const;
+                                         double& ingresos) const;
 
     /* ========================= RF-2.3 ========================= */
